requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
dnspython==2.4.2
python-nmap==0.7.1
python-whois==0.8.0
pysocks==1.7.1
stem==1.8.2
cryptography==41.0.7
lxml==4.9.3
validators==0.22.0
python-dotenv==1.0.0
